server/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.32
pydantic==2.8.2
pydantic-settings==2.3.4
email-validator==2.1.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
PyMySQL==1.1.1
alembic==1.13.2
python-multipart==0.0.9
Pillow==10.4.0
PyMuPDF==1.24.9
mammoth==1.8.0
APScheduler==3.10.4
httpx==0.27.0
python-dotenv==1.0.1